refactor(perl-dap): split process perl launch helpers (#0000)
